requirements.txt
requests
matplotlib
pandas
python-bitcoinrpc
awsglue
pyspark
numpy
pprint